requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
dateparser==0.7.0
gitdb2==2.0.3             # via gitpython
gitpython==2.1.8
python-dateutil==2.7.3    # via dateparser
pytz==2018.4              # via dateparser, tzlocal
regex==2018.2.21          # via dateparser
six==1.11.0               # via python-dateutil
smmap2==2.0.3             # via gitdb2
tzlocal==1.5.1            # via dateparser
